requirements.txt
django==3.0.2
djangorestframework==3.11.0
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
Pillow
pyyaml
django-allauth
django-tables2